Load zones from files outside the repo
In order to support dynamic updates, BIND needs to be able to write
to the zone file and create a .jnl record along side it. Since
the nix store can't be written to at run time, and because there
would be no way to sync dynamic updates with the repo, I have
moved them into a folder where they can be stored statefully.